Rewrite blake3 benchmarks with criterion
